% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus_function.R
\name{minus_function}
\alias{minus_function}
\title{My Minus Function}
\usage{
minus_function(x)
}
\arguments{
\item{x}{is a numeric value}
}
\value{
this returns x - 3
}
\description{
This function subtracts 3 numbers from an inputted value
}
\examples{
# An input of 5 should be 2
minus_function(5)

# An input of 7 should be 4
minus_function(7)

# An input of 10 should be 7
minus_function(10)
}
