% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_function.R
\name{sum_function}
\alias{sum_function}
\title{My Sum Function to Add Numbers}
\usage{
sum_function(x)
}
\arguments{
\item{x}{is a numeric value}
}
\value{
this returns x + 2
}
\description{
This function allows you to add 2 to any number inputted into the function
}
\examples{


# An input of 2 should be 4 
sum_function(x = 2)

#An input of 4 should be 6
sum_function(x = 4)

#An input of 10 should be 12
sum_function(x = 10)

}
