% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_two_function.R
\name{times_two_function}
\alias{times_two_function}
\title{My double function}
\usage{
times_two_function(x)
}
\arguments{
\item{x}{is a numeric value}
}
\value{
this returns x * 2
}
\description{
This function doubles any numeric value input into the function.
}
\examples{


# Output must be 10
times_two_function(5)

#Output must be 24
times_two_function(12)

#Output must be 48
times_two_function(24)
}
